refactor: renaming multer files into specific function file, and cleaning duplicate obsolete profile_pic_url
